Make requeuingStrategy mutable
